Cleanup TaskItems and collect similar container code
